Extract SectionActions and isolate editor store
Wrap resolved component in a functional setter to prevent it being
interpreted as a React state updater. Remove the package.json "prepare"
script (husky install). Regenerate lockfiles (add vitest coverage deps
and platform-specific @next/swc packages)
